Parte UI

+++ lib/models/Corso.dart
@@ -1,5 +1,4 @@
 import 'dart:convert';
-import 'dart:ffi';
 import 'package:firebase_database/firebase_database.dart';
 import 'package:progetto_camilloni_tiseni_giri/models/Documento.dart';
 import 'package:progetto_camilloni_tiseni_giri/models/Lezione.dart';
@@ -26,7 +25,7 @@ class Corso {
   List<Documento> dispense;
   String immagine;
   List<Lezione> lezioni;
-  List<Float> recensioni;
+  List<double> recensioni;
   String titolo;
 
   factory Corso.fromJson(Map<String, dynamic> json) => Corso(
@@ -40,7 +39,7 @@ class Corso {
             List<Lezione>.from(json["lezioni"].map((x) => Lezione.fromJson(x))),
         titolo: json["titolo"],
     recensioni:
-    List<Float>.from(json["recensioni"].map((x) => x)),
+    List<double>.from(json["recensioni"].map((x) => x)),
       );
 
   Map<String, dynamic> toJson() => {
